requirements.txt
Click==7.0
coverage==4.5.3
Flask==1.0.3
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.10.3
SQLAlchemy==1.3.4
Werkzeug==0.15.4
